requirements.txt
# Production dependencies for signage project
# All versions pinned exactly for reproducible builds
# Last updated: 2025-11-28

# Core dependencies
requests==2.32.5
python-dotenv==1.2.1
pytz==2025.2
Jinja2==3.1.6
pydantic==2.12.4
pydantic-settings==2.11.0
PyYAML==6.0.3
croniter==6.0.0
python-dateutil==2.9.0.post0
six==1.17.0

# Browser automation (Playwright for HTML rendering)
playwright==1.56.0

# HTTP client with retry support
urllib3==2.5.0

# Samsung TV integration
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
pillow==11.3.0
propcache==0.4.1
py3rijndael==0.3.3
pycparser==2.23
typing_extensions==4.15.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
samsungtvws @ git+https://github.com/NickWaterton/samsung-tv-ws-api.git@7e7766807da0b048e441eef86e7130a5f5030ed9

# Pydantic core (required by pydantic)
pydantic_core==2.41.5
annotated-types==0.7.0

# Template rendering dependencies
MarkupSafe==3.0.3

# Playwright dependencies
greenlet==3.2.4
pyee==13.0.0
psutil==7.1.3
